pyproject.toml
[project]
name = "utcm-exporter"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Catalin Cirstoiu", email = "Catalin.Cirstoiu@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "msal>=1.34.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
utcm-exporter = "utcm_exporter:main"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"
